refactor: move plane creation to common place
