+++ tools/dark_viewer/src/scenes/bin_obj_viewer.rs
@@ -1,11 +1,14 @@
 #![allow(unused_imports)]
 
-use super::{ToolScene, render_helpers::build_model_scene_with_debug_skeletons};
+use super::{
+    ToolScene,
+    render_helpers::{build_model_scene_with_debug_skeletons, create_ground_plane},
+};
 use cgmath::{Deg, Matrix4, Quaternion, Rad, vec3};
-use dark::importers::{MODELS_IMPORTER, TEXTURE_IMPORTER};
+use dark::importers::MODELS_IMPORTER;
 use dark::motion::AnimationPlayer;
 use engine::assets::asset_cache::AssetCache;
-use engine::scene::{Scene, SceneObject, color_material, basic_material, create_plane_with_uv_scale};
+use engine::scene::{Scene, SceneObject, color_material};
 use std::time::Duration;
 
 pub struct BinObjViewerScene {
@@ -61,16 +64,7 @@ impl ToolScene for BinObjViewerScene {
         let mut turret_scene_objects = turret.to_animated_scene_objects(&self.animation_player);
 
         // Add ground plane
-        let grid_texture = asset_cache.get(&TEXTURE_IMPORTER, "grid.png");
-        let texture_trait: std::rc::Rc<dyn engine::texture::TextureTrait> = grid_texture;
-        let ground_material = basic_material::create(texture_trait, 1.0, 0.5); // 100% emissivity, 50% transparency
-        let ground_plane = SceneObject::new(ground_material, Box::new(create_plane_with_uv_scale(10.0)));
-
-        // Scale the ground plane to be larger (10x10 units)
-        let scale_transform = Matrix4::from_scale(10.0);
-        let mut ground_plane_scaled = ground_plane;
-        ground_plane_scaled.set_transform(scale_transform);
-        turret_scene_objects.push(ground_plane_scaled);
+        turret_scene_objects.push(create_ground_plane(asset_cache));
 
         build_model_scene_with_debug_skeletons(
             turret.as_ref(),
