+++ tools/dark_viewer/src/scenes/glb_viewer.rs
@@ -1,10 +1,10 @@
 #![allow(unused_imports)]
 
-use super::ToolScene;
+use super::{ToolScene, render_helpers::create_ground_plane};
 use cgmath::{Deg, Matrix4, Quaternion, Rad, SquareMatrix, vec3};
-use dark::importers::{GLB_MODELS_IMPORTER, TEXTURE_IMPORTER};
+use dark::importers::GLB_MODELS_IMPORTER;
 use engine::assets::asset_cache::AssetCache;
-use engine::scene::{Scene, SceneObject, color_material, basic_material, create_plane_with_uv_scale};
+use engine::scene::{Scene, SceneObject, color_material};
 use std::rc::Rc;
 use std::time::Duration;
 
@@ -49,16 +49,7 @@ impl ToolScene for GlbViewerScene {
         }
 
         // Add ground plane
-        let grid_texture = asset_cache.get(&TEXTURE_IMPORTER, "grid.png");
-        let texture_trait: std::rc::Rc<dyn engine::texture::TextureTrait> = grid_texture;
-        let ground_material = basic_material::create(texture_trait, 1.0, 0.5); // 100% emissivity, 50% transparency
-        let ground_plane = SceneObject::new(ground_material, Box::new(create_plane_with_uv_scale(10.0)));
-
-        // Scale the ground plane to be larger (10x10 units)
-        let ground_scale_transform = Matrix4::from_scale(10.0);
-        let mut ground_plane_scaled = ground_plane;
-        ground_plane_scaled.set_transform(ground_scale_transform);
-        scene_objects.push(ground_plane_scaled);
+        scene_objects.push(create_ground_plane(asset_cache));
 
         // Add debug skeleton visualization if requested
         if self.debug_skeletons && model.has_skeleton() {
